Testing for all the local database code.
A test I wrote in 3.7 crashed in 2.7 (but not pypy) having something to do with the CFFI allocations; I've disabled pre-allocation for now until I can look more closely. (But hopefully one of two things will happen: either more people will use persistent caches and get the contiguous benefits that way, or a different LRU implementation such as sqlite will replace the custom one and we can drop CFFI altogether.)
